Refactor hearing test welcome page: modularize components and improve layout with new widgets for results display
